// models/models.tsx
export interface IMetaDetails {
  title: string;
  description: string;
  keyword: string;
  ogTitle: string;
}
// ...
export interface IHomePageSliderSpec {
  spec1: string;
  spec2: string;
}

export interface IPowerProducts {
  id: number;
  showInMainMenu: boolean;
  homePageSliderSpec: IHomePageSliderSpec;
  specs: IKeyValue[];
  price: string;
  offerPrice: string;
  type: string;
  category: string;
  slug: string;
  title: string;
  subTitle: string;
  locale: string;
  localizations: ILocalization[];
  megaMenuThumb: IImage;
  banner: ITopBanner[];
  description: string;
  detailItems: DetailItem[];
  buildLink: string;
  accordian: IAccordian[];
  displayImage: IImage;
  mainSpecifications: IThreeColumnData[];
  metaDetails?: IMetaDetails;
  ecomProductId: string;
}

export interface IThreeColumnData {
  col1: string;
  col2: string;
  col3: string;
  col4: string;
}

export interface IAccordian {
  title: string;
  headerWithSpec: IHeaderWithSpec[];
}

export interface IHeaderWithSpec {
  label: string;
  value: string;
}

export interface IConvenience {
  title: string;
  description: string;
  image: IImage;
}

export interface ITrims {
  title: string;
  subTitle: string;
  price: string;
  infoText: string;
  spec: ISpecText[];
  image: IImage;
}

export interface ISpecText {
  spefcItem: string;
}

export interface IColorVariant {
  title: string;
  colorIcon: IImage;
  image: IImage;
}
export interface DetailItem {
  title: string;
  subtitle?: any;
  items: Item[];
}

export interface Item {
  id: number;
  title: string;
  description: string;
  image: IImage;
}

export interface IHighlights {
  id: number;
  title: string;
  subtitle: string;
  items: Item[];
}

export interface IKeyValue {
  label: string;
  value: string;
}

export interface IImage {
  id: number;
  name: string;
  url: string;
}

export interface IDesign {
  id: number;
  title: string;
  items: IImageWithDescription[];
}

export interface IImageWithDescription {
  id: number;
  title: string;
  description: string;
  image: IImage;
}

export interface ILocalization {
  locale: string;
  id: string;
  slug: string;
}
// ...
export interface ITopBanner {
  title: string;
  description: string;
  type: string;
  image: {
    url: string;
    formats: IimageFormat;
  };
  automobile: IRelationData;
  motorcycle: IRelationData;
  marine: IRelationData;
  power_product: IRelationData;
  button: IButton[];
  relativeUrl: string;
  absoluteUrl: string;
}

export interface IimageFormat {
  small: { url: string };
}

export interface IRelationData {
  title: string;
  slug: string;
  locale: string;
  localizations: ILocalization[];
}
export interface IButton {
  color: "light_color" | "theme_color";
  text: string;
  relativeUrl: string;
  absoluteUrl: string;
  automobile: IRelationData;
  motorcycle: IRelationData;
  marine: IRelationData;
  power_product: IRelationData;
}
export interface IWelcomeText {
  title: string;
  description: string;
}
export interface IShoppingTool {
  title: string;
  description: string;
  items: IShoppingToolItem[];
}
export interface IShoppingToolItem extends ITopBanner {
  link: string;
}
export interface ISocialMedia {
  title: string;
  items: ISocialMediaItem[];
}
export interface ISocialMediaItem {
  url: string;
  iconClass: string;
  position: string;
  image: {
    url: string;
  };
}
// ...
export interface ILinkItem {
  label: string;
  relativeUrl?: string;
  absoluteUrl?: string;
}

export interface IFooterMenu {
  title: string;
  items: ILinkItem[];
  isDisable?: boolean;
}
export interface IFooter {
  title: string;
  copyRight: string;
  menuItems: IFooterMenu[];
  socialMedia: ISocialMedia[];
  serviceLinks: IFooterMenu[];
  connectWithUs: IFooterMenu[];
  extraLinks: IFooterMenu[];
}
// ...
export interface IAutoMobileCategoryItem {
  title: string;
  subTitle: string;
  link: string;
  image: {
    url: string;
  };
}
export interface IPowerProductMenuLandingItem {
  title: string;
  slug: string;
  price: string;
  image: IImage;
  category?: string;
  type?: string;
}
export interface IMotorCylceLandingPage {
  title: string;
  banner: ITopBanner[];
  categoryList: IAutoMobileCategoryItem[];
  metaDetails?: IMetaDetails;
  slider: IHomepageSlider[];
}

export interface IPowerProductMenuLandingPage {
  title: string;
  banner: ITopBanner[];
  categoryList: IAutoMobileCategoryItem[];
  metaDetails?: IMetaDetails;
  items: IPowerProductMenuLandingItem[];
}

export interface IMarineMenuLandingPage {
  title: string;
  banner: ITopBanner[];
  categoryList: IAutoMobileCategoryItem[];
  metaDetails?: IMetaDetails;
  items: IPowerProductMenuLandingItem[];
}